Module kanidmd_core::https::middleware

Structs

CacheableMiddleware

NoCacheMiddleware

StaticContentMiddleware

Sets Cache-Control headers on static content endpoints

StrictResponseMiddleware

Adds the following headers to responses

UIContentSecurityPolicyResponseMiddleware

This tide MiddleWare adds headers like Content-Security-Policy and similar families. If it keeps adding more things then probably rename the middleware :)

VersionHeaderMiddleware

Functions

compression_content_type_checker

This is for the tide_compression middleware so that we only compress certain content types.